Update Wordle3.java
The 5 for loops can each be terminated early.
The word list is sorted lexicographically.
If the words are ABCD and E, and A > B > C > D > E. Once |A| has 24 or fewer bits, then so do B,C,D and E. So the maximum of any union can only be 24.
This is known and implemented in everyone's version.
But the other loops can be optimized similarly. Once |B| has 19 or fewer bits, then so does the maximum union of B, C, D and E. So B can be exited early. Likewise, no need to check C once it has 14 or few bits; D must have 9 or more and E must have 4 or more. The case with E is irrelavent because all elements have 5 bits.
The indices of where these the transitions from say, 25 to 24 bits occurs is found using the insertion point returned (but not found) of the Arrays.binarySearch function. A search for 1 << 24 returns a negative number (indicating it was not found as expected) but its insertion point is the first number less than that.
BinarySearch only works if the array is sorted ascendingly, but we want our array sorted descendingly. So we adjust the insertion points by subtracting them from the length and reverse the array after we've done the searches.

The optimization on the parallel loop of A is interesting. The cost of a loop around A=words[0] is much higher than A=words[i] due to the time complexity being on the order of (N-i) choose 4. So if we're looking to divide up the cost of the first loop only, we shouldn't do it with an arithmetic sequence. But either a round robin or something like size^(i/(k+1)), where k is the number of threads.
Overall the algorithm takes (n choose 5) * p^4, where p is the probability that two words chosen at random will not contain any mutual letters. I calculated this to be about 0.28.
I calculated this by taking a frequency table of all the english letters and putting them in a spreadsheet. For each letter I calculated:
Probability of Letter, p (for E => 0.12, T => .09 ...)
Probability of not letter 1-p
5 letters all not (assuming independence which is not valid but let's go with it!) (1-p)^5
Word containing at least 1 letter, (complement of above)
2 words containing at least 1 letter, (square of above)
2 words don't contain at least 1 letter, (complement of above)
Now we're looking for the probability that 
2 words don't contain at least 1 a AND
2 words don't contain at least 1 b AND
2 words don't contain at least 1 c AND...
So we multiply them all out to get 0.28
This isn't too far off from the number of cliques of various sizes calculated: 
number of cliques of size(0,1,2,3,4,5)
1, 5977, 3213696, 95866204, 26133319, 538
The 0th term is included as the empty set.
p is 3213696 / (5977 choose 2) = 0.18.

Incidentally the "Parker Algorithm" is soooo slow not because he's using Sets instead of Ints, but because he has a pairs vs pairs loop which takes
( n choose 2 * p) choose 2 = 5.16391938336 × 10^12

If we include the early termination of the A loop we get number of cliques of size i =
1, 460, 554861, 29450403, 14756251, 538

I also used groupingBy to handle anagrams, as it's the perfect tool for the job. GroupingBy uses a function String=>Int which he have. When Ints "collide", they are grouped into a list. The result is a mapping from Ints to List[String], as desired. The keys (ints) of this map is easily extracted then sorted to be the list of words to work with. When we go to print a word, all the anagrams are easily accessed by map[word] which returns the list.